Revert "Merge tiny Chunk mixin into MixinChunk"
This reverts commit 1010085d52496bcca8d44c212cd843dfdfec77ff.
