+++ src/main/java/com/cardinalstar/cubicchunks/mixin/early/client/MixinChunk_Cubes.java
@@ -20,61 +20,69 @@
  */
 package com.cardinalstar.cubicchunks.mixin.early.client;
 
-import java.util.Collection;
-import java.util.Collections;
-
 import javax.annotation.ParametersAreNonnullByDefault;
 
-import net.minecraft.world.World;
 import net.minecraft.world.chunk.Chunk;
-import net.minecraft.world.chunk.EmptyChunk;
 import net.minecraft.world.chunk.storage.ExtendedBlockStorage;
 
+import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
+import com.cardinalstar.cubicchunks.api.IColumn;
+import com.cardinalstar.cubicchunks.api.IHeightMap;
 import com.cardinalstar.cubicchunks.world.column.CubeMap;
-import com.cardinalstar.cubicchunks.world.cube.BlankCube;
 import com.cardinalstar.cubicchunks.world.cube.Cube;
 
+/**
+ * Modifies vanilla code in Chunk to use Cubes. Client side only.
+ */
 @ParametersAreNonnullByDefault
-@Mixin(EmptyChunk.class)
-public abstract class MixinEmptyChunk extends MixinChunk_Cubes {
+@Mixin(Chunk.class)
+public abstract class MixinChunk_Cubes implements IColumn {
 
-    private Cube blankCube;
+    protected MixinChunk_Cubes() {}
 
-    @Inject(method = "<init>", at = @At(value = "RETURN"))
-    private void cubicChunkColumn_construct(World worldIn, int x, int z, CallbackInfo cbi) {
-        blankCube = new BlankCube((Chunk) (Object) this);
+    @Shadow
+    public abstract ExtendedBlockStorage[] getBlockStorageArray();
+
+    @Shadow
+    @Final
+    private ExtendedBlockStorage[] storageArrays;
+    /*
+     * WARNING: WHEN YOU RENAME ANY OF THESE 3 FIELDS RENAME CORRESPONDING
+     * FIELDS IN "cubicchunks.mixin.early.common.MixinChunk_Cubes" and
+     * "cubicchunks.mixin.early.common.MixinChunk_Columns".
+     */
+    private CubeMap cubeMap;
+    private IHeightMap opacityIndex;
+    private Cube cachedCube; // todo: make it always nonnull using BlankCube
+
+    private boolean isColumn = false;
+
+    // ==============================================
+    // generateHeightMap
+    // ==============================================
+
+    @Inject(method = "generateHeightMap", at = @At(value = "HEAD"), cancellable = true)
+    private void generateHeightMap_CubicChunks_Cancel(CallbackInfo cbi) {
+        if (isColumn) {
+            cbi.cancel();
+        }
     }
 
-    @Override
-    public Cube getCube(int cubeY) {
-        return blankCube;
-    }
+    // ==============================================
+    // fillChunk
+    // ==============================================
 
-    @Override
-    public Cube removeCube(int cubeY) {
-        return blankCube;
-    }
-
-    @Override
-    public void addCube(Cube cube) {}
-
-    @Override
-    public Collection<Cube> getLoadedCubes() {
-        return Collections.emptySet();
-    }
-
-    @Override
-    public ExtendedBlockStorage[] getTickableStorages() {
-        return CubeMap.ZERO_LEN_EBS_ARRAY;
-    }
-
-    @Override
-    public Iterable<Cube> getLoadedCubes(int startY, int endY) {
-        return Collections.emptySet();
+    @Inject(method = "fillChunk", at = @At(value = "HEAD"))
+    private void fillChunk_CubicChunks_NotSupported(byte[] p_76607_1_, int p_76607_2_, int p_76607_3_,
+        boolean p_76607_4_, CallbackInfo ci) {
+        if (false) if (isColumn) {
+            throw new UnsupportedOperationException("setting storage arrays it not supported with cubic chunks");
+        }
     }
 }
