+++ src/main/java/com/cardinalstar/cubicchunks/mixin/early/common/MixinChunk.java
@@ -277,29 +277,6 @@ public abstract class MixinChunk implements IColumn, IColumnInternal {
     }
 
     // ==============================================
-    // generateHeightMap
-    // ==============================================
-
-    @Inject(method = "generateHeightMap", at = @At(value = "HEAD"), cancellable = true)
-    private void generateHeightMap_CubicChunks_Cancel(CallbackInfo cbi) {
-        if (isColumn) {
-            cbi.cancel();
-        }
-    }
-
-    // ==============================================
-    // fillChunk
-    // ==============================================
-
-    @Inject(method = "fillChunk", at = @At(value = "HEAD"))
-    private void fillChunk_CubicChunks_NotSupported(byte[] p_76607_1_, int p_76607_2_, int p_76607_3_,
-        boolean p_76607_4_, CallbackInfo ci) {
-        if (false) if (isColumn) {
-            throw new UnsupportedOperationException("setting storage arrays it not supported with cubic chunks");
-        }
-    }
-
-    // ==============================================
     // relightBlock
     // ==============================================
